Allow unsafe access to LR pool to avoid locks
Unsafe access is okay if no other thread accesses the same core at the
same time. This is the case in STS.

The map of vector of LRs was replaced with a vector of vectors to make
unsafe access possible. The user now, however, has to provide the set
of cores before using the pool.